chore: move more things around in version 1 api
