+++ roborock/containers.py
@@ -829,3 +829,28 @@ class RoborockCategoryDetail(RoborockBase):
 @dataclass
 class ProductResponse(RoborockBase):
     category_detail_list: list[RoborockCategoryDetail]
+
+
+@dataclass
+class DyadProductInfo(RoborockBase):
+    sn: str
+    ssid: str
+    timezone: str
+    posix_timezone: str
+    ip: str
+    mac: str
+    oba: dict
+
+
+@dataclass
+class DyadSndState(RoborockBase):
+    sid_in_use: int
+    sid_version: int
+    location: str
+    bom: str
+    language: str
+
+
+@dataclass
+class DyadOtaNfo(RoborockBase):
+    mqttOtaData: dict
