+++ roborock/version_1_apis/roborock_local_client_v1.py
@@ -1,6 +1,9 @@
+import asyncio
+
 from roborock.local_api import RoborockLocalClient
 
-from .. import DeviceData, RoborockCommand
+from .. import CommandVacuumError, DeviceData, RoborockCommand, RoborockException
+from ..protocol import MessageParser
 from ..roborock_message import MessageRetry, RoborockMessage, RoborockMessageProtocol
 from .roborock_client_v1 import COMMANDS_SECURED, RoborockClientV1
 
@@ -8,7 +11,7 @@ from .roborock_client_v1 import COMMANDS_SECURED, RoborockClientV1
 class RoborockLocalClientV1(RoborockLocalClient, RoborockClientV1):
     def __init__(self, device_data: DeviceData, queue_timeout: int = 4):
         RoborockLocalClient.__init__(self, device_data, queue_timeout)
-        RoborockClientV1.__init__(self, device_data, self.cache, self._logger, "abc")
+        RoborockClientV1.__init__(self, device_data, self._logger, "abc")
 
     def build_roborock_message(
         self, method: RoborockCommand | str, params: list | dict | int | None = None
@@ -30,3 +33,40 @@ class RoborockLocalClientV1(RoborockLocalClient, RoborockClientV1):
     ):
         roborock_message = self.build_roborock_message(method, params)
         return await self.send_message(roborock_message)
+
+    async def send_message(self, roborock_message: RoborockMessage):
+        await self.validate_connection()
+        method = roborock_message.get_method()
+        params = roborock_message.get_params()
+        request_id: int | None
+        if not method or not method.startswith("get"):
+            request_id = roborock_message.seq
+            response_protocol = request_id + 1
+        else:
+            request_id = roborock_message.get_request_id()
+            response_protocol = RoborockMessageProtocol.GENERAL_REQUEST
+        if request_id is None:
+            raise RoborockException(f"Failed build message {roborock_message}")
+        local_key = self.device_info.device.local_key
+        msg = MessageParser.build(roborock_message, local_key=local_key)
+        if method:
+            self._logger.debug(f"id={request_id} Requesting method {method} with {params}")
+        # Send the command to the Roborock device
+        async_response = asyncio.ensure_future(self._async_response(request_id, response_protocol))
+        self._send_msg_raw(msg)
+        (response, err) = await async_response
+        self._diagnostic_data[method if method is not None else "unknown"] = {
+            "params": roborock_message.get_params(),
+            "response": response,
+            "error": err,
+        }
+        if err:
+            raise CommandVacuumError(method, err) from err
+        if roborock_message.protocol == RoborockMessageProtocol.GENERAL_REQUEST:
+            self._logger.debug(f"id={request_id} Response from method {roborock_message.get_method()}: {response}")
+        if response == "retry":
+            retry_id = roborock_message.get_retry_id()
+            return self.send_command(
+                RoborockCommand.RETRY_REQUEST, {"retry_id": retry_id, "retry_count": 8, "method": method}
+            )
+        return response
