+++ roborock/version_1_apis/roborock_mqtt_client_v1.py
@@ -8,7 +8,10 @@ from roborock.cloud_api import RoborockMqttClient
 from ..containers import DeviceData, UserData
 from ..exceptions import CommandVacuumError, RoborockException
 from ..protocol import MessageParser, Utils
-from ..roborock_message import RoborockMessage, RoborockMessageProtocol
+from ..roborock_message import (
+    RoborockMessage,
+    RoborockMessageProtocol,
+)
 from ..roborock_typing import RoborockCommand
 from .roborock_client_v1 import COMMANDS_SECURED, RoborockClientV1
 
@@ -21,7 +24,7 @@ class RoborockMqttClientV1(RoborockMqttClient, RoborockClientV1):
         endpoint = base64.b64encode(Utils.md5(rriot.k.encode())[8:14]).decode()
 
         RoborockMqttClient.__init__(self, user_data, device_info, queue_timeout)
-        RoborockClientV1.__init__(self, device_info, self.cache, self._logger, endpoint)
+        RoborockClientV1.__init__(self, device_info, self._logger, endpoint)
 
     def _send_msg_raw(self, msg: bytes) -> None:
         info = self.publish(f"rr/m/i/{self._mqtt_user}/{self._hashed_user}/{self.device_info.device.duid}", msg)
